Compact common tests and prevent comiler warnings on Arrow functions
