+++ common/cpp/test/fletcher/test-common.cc
@@ -20,80 +20,7 @@
 
 #include "fletcher/test_schemas.h"
 #include "fletcher/test_recordbatches.h"
-
-TEST(Common, appendExpectedBuffersFromField) {
-  // List of uint8's
-  auto schema = fletcher::GetListUint8Schema();
-  std::vector<std::string> bufs;
-  fletcher::AppendExpectedBuffersFromField(&bufs, schema->field(0));
-  ASSERT_EQ(bufs[0], "list_offsets");
-  ASSERT_EQ(bufs[1], "number_values");
-
-  schema = fletcher::GetStringReadSchema();
-  // String is essentially a list of non-nullable utf8 bytes
-  std::vector<std::string> bufs2;
-  fletcher::AppendExpectedBuffersFromField(&bufs2, schema->field(0));
-  ASSERT_EQ(bufs2[0], "Name_offsets");
-  ASSERT_EQ(bufs2[1], "Name_values");
-}
-
-TEST(Common, RecordBatchFileRoundTrip) {
-  auto wrb = fletcher::getStringRB();
-  fletcher::WriteRecordBatchToFile(wrb, "test-common.rb");
-  auto rrb = fletcher::ReadRecordBatchFromFile("test-common.rb", wrb->schema());
-  ASSERT_TRUE(wrb->Equals(*rrb));
-}
-
-TEST(Common, flattenArrayBuffers_string) {
-  // Test flattening of array buffers with field
-  auto rb = fletcher::getStringRB();
-  auto sa = std::dynamic_pointer_cast<arrow::StringArray>(rb->column(0));
-
-  std::vector<arrow::Buffer*> buffers;
-  fletcher::FlattenArrayBuffers(&buffers, rb->column(0), rb->schema()->field(0));
-  // First buffer should be offsets buffer
-  ASSERT_EQ(buffers[0], sa->value_offsets().get());
-  // Second buffer should be values buffer
-  ASSERT_EQ(buffers[1], sa->value_data().get());
-}
-
-TEST(Common, flattenArrayBuffers_string_noField) {
-  // Test flattening of array buffers without field
-  auto rb = fletcher::getStringRB();
-  auto sa = std::dynamic_pointer_cast<arrow::StringArray>(rb->column(0));
-
-  std::vector<arrow::Buffer*> buffers;
-  fletcher::FlattenArrayBuffers(&buffers, rb->column(0));
-  // First buffer should be offsets buffer
-  ASSERT_EQ(buffers[0], sa->value_offsets().get());
-  // Second buffer should be values buffer
-  ASSERT_EQ(buffers[1], sa->value_data().get());
-}
-
-TEST(Common, flattenArrayBuffers_list) {
-  // Test flattening of array buffers without field
-  auto rb = fletcher::getListUint8RB();
-  auto la = std::dynamic_pointer_cast<arrow::ListArray>(rb->column(0));
-  auto va = std::dynamic_pointer_cast<arrow::UInt8Array>(la->values());
-
-  std::vector<arrow::Buffer*> buffers;
-  fletcher::FlattenArrayBuffers(&buffers, rb->column(0), rb->schema()->field(0));
-  // First buffer should be offsets buffer
-  ASSERT_EQ(buffers[0], la->value_offsets().get());
-  // Second buffer should be values buffer
-  ASSERT_EQ(buffers[1], va->values().get());
-}
-
-TEST(Common, AppendMetaRequired) {
-  auto schema = fletcher::GetPrimReadSchema();
-  ASSERT_NE(schema->metadata(), nullptr);
-  std::unordered_map<std::string, std::string> map;
-  schema->metadata()->ToUnorderedMap(&map);
-  ASSERT_TRUE(map.count("fletcher_name") > 0);
-  ASSERT_TRUE(map.count("fletcher_mode") > 0);
-  ASSERT_TRUE(map.at("fletcher_name") == "PrimRead");
-  ASSERT_TRUE(map.at("fletcher_mode") == "read");
-}
+#include "fletcher/test_common.h"
 
 int main(int argc, char **argv) {
   ::testing::InitGoogleTest(&argc, argv);
